Implement partial expansion of alternative patterns.
This allows glob with patterns including {} to be run more efficiently. Most glob patterns using {} only involves one filename component (no '/' inside the path), and does not need to be expanded before globbing. This implements expansion only when resolving filenames (using fnmatch)

This also reduces the asymptotic complexity of glob when {} only involves one filename component: from an algorithm requiring O(mn) calls to ruby_glob0, where m is the number of options in one set, and n is the number of sets in the filename, this makes globbing O(1), with a corresponding decrease in the amount of IO operations performed by Ruby.